Changed type of weight to double; Added checked arg. in abstract class Item

+++ src/some/simple/items/Ball.java
@@ -5,7 +5,7 @@ import some.Shape;
 
 public class Ball extends Item {
 
-    public Ball(String name, int weight, int size, String color) {
+    public Ball(String name, double weight, int size, String color) {
         super(name, Shape.ROUND, weight, size, color);
     }
 }
